+++ src/server/index.ts
@@ -1,4 +1,5 @@
 export * from './BaseMcpServer.js';
+export * from './EmbeddableMcpServer.js';
 export * from './IHttpApplication.js';
 export * from './SseServer.js';
 export * from './StdioServer.js';